Feat: Add Delete or Update reconfirm

+++ src/containers/AdminPage/DishSetting.js
@@ -73,14 +73,16 @@ const DishSetting = () => {
     });
   }
   function deleteDishBtn(dishNumber) {
-    DishService.deleteDish(dishNumber)
-    .then((response) => {
-      alert("해당 요리 데이터가 삭제되었습니다.")
-      window.location.reload()
-    })
-    .catch((error) => {
-      console.log(error.response);
-    });
+    if (window.confirm("정말 삭제하시겠습니까?")) {
+      DishService.deleteDish(dishNumber)
+      .then((response) => {
+        alert("해당 요리 데이터가 삭제되었습니다.")
+        window.location.reload()
+      })
+      .catch((error) => {
+        console.log(error.response);
+      });
+    }
   }
   function SearchOutput({dishNumber, dishName}){
     console.log(dishNumber + ", " + dishName)
